\name{QR}
\alias{QR}
\docType{data}
\title{Wind power scenarios generated by Quantile Regression with Gaussian copula approach}
\description{This data set provides 24-hour wind power scenarios and observations for 345 instances (days)}
\usage{data("QR")}
\format{A list containing wind power scenarios for 345 days, where each day includes 27 scenarios and one observation of dimension 24, representing hourly values.
}

\references{
[2]D. Sari, S.Ryan. Reliability of wind power scenarios and stochastic unit commitment cost. Under review. 
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
data(QR)
qr_ranks <- MTDrh(QR$scen,QR$obs,NULL,FALSE,FALSE)
}
\keyword{datasets}
